fix: score the reference backend in float32 and document the noise floor
A GPU smoke run showed the reference and vLLM paths disagreeing by 0.13 mean
log-prob per token, which would be large enough to swamp the credit signal.
Measuring the pairwise gaps on Qwen3-0.6B locates the cause:

  HF bfloat16 vs HF float32          0.236
  vLLM vs HF bfloat16                0.213
  vLLM vs HF float32                 0.105
  vLLM prefix caching on vs off      0.075

The reference was the inaccurate side, not vLLM. Scoring in bfloat16 carries
more numerical noise than the disagreement being adjudicated, so the HF backend
now defaults to float32, exposed as --reference-dtype.

The remaining figures are recorded in the README because they bound what the
recipe can resolve: a credit difference of a few tenths of a nat per token on a
small model is not distinguishable from arithmetic, and prefix caching, which
is what makes the ablation variants affordable, is itself not numerically free.

+++ examples/scoring_credit/backends.py
@@ -100,10 +100,10 @@ class VLLMBackend:
 class HFBackend:
     """Reference span scoring with a plain Hugging Face forward pass."""
 
-    # batch_size defaults to 1 so the reference path involves no padding at all.
-    # This backend exists to adjudicate disagreements, so it is worth more as an
-    # obviously-correct implementation than as a fast one.
-    def __init__(self, model: str, device: str = "cuda", dtype: str = "bfloat16", batch_size: int = 1):
+    # Defaults chosen so the reference is trustworthy rather than fast: float32
+    # because bfloat16 log-prob scoring carries more numerical noise than the gap
+    # being adjudicated, and batch_size 1 so no padding is involved at all.
+    def __init__(self, model: str, device: str = "cuda", dtype: str = "float32", batch_size: int = 1):
         import torch
         from transformers import AutoModelForCausalLM
